Fix progress bar bug when downloading multiple times shaders and / or firmware, move some logic to the right place
